pyproject.toml
[tool.poetry]
name = "robot_for_ig_followers"
version = "0.1.0"
description = ""
authors = ["FabricioPatrocinio <fabricioschiffer@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
python-dotenv = "^0.19.2"
pydantic = "^1.9.0"
alive-progress = "^2.3.1"
playwright = "^1.19.1"
typer = "^0.4.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
ipdb = "^0.13.9"
termcolor = "^1.1.0"
ipython = "^8.1.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
